webapp/requirements.txt
streamlit==1.11.1
tensorflow==2.9.1
numpy==1.23.1
pandas==1.4.2
matplotlib==3.5.3
Pillow==9.2.0





